+++ examples/shared/src/examples/MapView/Compass.tsx
@@ -1,10 +1,8 @@
 import { Camera, MapView } from "@maplibre/maplibre-react-native";
 
-import { sheet } from "../../styles/sheet";
-
 export function Compass() {
   return (
-    <MapView style={sheet.matchParent} compass>
+    <MapView compass>
       <Camera bearing={21} />
     </MapView>
   );