+++ examples/shared/src/examples/MapView/SetTintColor.tsx
@@ -2,7 +2,6 @@ import { Camera, MapView, UserLocation } from "@maplibre/maplibre-react-native";
 import { useState } from "react";
 
 import { TabBarView } from "../../components/TabBarView";
-import { sheet } from "../../styles/sheet";
 
 const OPTIONS = ["red", "yellow", "green"].map((data) => ({
   label: data,
@@ -19,7 +18,7 @@ export function SetTintColor() {
         setTintColor(data);
       }}
     >
-      <MapView style={sheet.matchParent} tintColor={tintColor}>
+      <MapView tintColor={tintColor}>
         <Camera zoom={6} trackUserLocation="heading" />
 
         <UserLocation renderMode="native" androidRenderMode="compass" />