+++ examples/shared/src/examples/UserLocation/UserLocationUpdate.tsx
@@ -1,6 +1,6 @@
 import {
   Camera,
-  type Location,
+  type GeolocationPosition,
   MapView,
   UserLocation,
 } from "@maplibre/maplibre-react-native";
@@ -8,14 +8,13 @@ import { useState } from "react";
 import { Text } from "react-native";
 
 import { Bubble } from "../../components/Bubble";
-import { sheet } from "../../styles/sheet";
 
 export function UserLocationUpdate() {
-  const [location, setLocation] = useState<Location>();
+  const [location, setLocation] = useState<GeolocationPosition>();
 
   return (
     <>
-      <MapView style={sheet.matchParent}>
+      <MapView>
         <UserLocation onUpdate={(newLocation) => setLocation(newLocation)} />
         <Camera trackUserLocation="default" zoom={16} />
       </MapView>
@@ -26,9 +25,10 @@ export function UserLocationUpdate() {
             <Text>Timestamp: {location.timestamp}</Text>
             <Text>Longitude: {location.coords.longitude}</Text>
             <Text>Latitude: {location.coords.latitude}</Text>
-            <Text>Altitude: {location.coords.altitude}</Text>
-            <Text>Heading: {location.coords.heading}</Text>
             <Text>Accuracy: {location.coords.accuracy}</Text>
+            <Text>Altitude: {location.coords.altitude}</Text>
+            <Text>Altitude Accuracy: {location.coords.altitudeAccuracy}</Text>
+            <Text>Heading: {location.coords.heading}</Text>
             <Text>Speed: {location.coords.speed}</Text>
           </>
         )}